requirements.txt
PyMuPDF==1.26.3
python-docx==1.2.0
google_generativeai==0.8.5
faiss-cpu==1.11.0
joblib==1.5.1
scikit-learn==1.7.1
streamlit==1.47.0